agri-connect/ml-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
Pillow==10.4.0
numpy==2.1.1
opencv-python==4.10.0.84
torch==2.3.1
transformers==4.44.2
rembg==2.0.56
